Move resources module up a level
