+++ src/napari_imagej/resources/__init__.py
