+++ src/napari_imagej/resources/imagej2-16x16-flat-disabled.png
